+++ Sources/StreamChatSwiftUI/ChatChannelList/SearchResultsView.swift
@@ -158,7 +158,7 @@ struct SearchResultItem<Factory: ViewFactory, ChannelDestination: View>: View {
 
     private var timestampText: String {
         if let lastMessageAt = searchResult.channel.lastMessageAt {
-            return utils.dateFormatter.string(from: lastMessageAt)
+            return utils.channelListDateFormatter.string(from: lastMessageAt)
         } else {
             return ""
         }